+++ controllers/jobs.js
@@ -4,9 +4,11 @@ const deleteData = require("./deleteData");
 const updateData = require("./updateData");
 const conn = require("../services/db");
 const auth = require("./authentication");
+const user = require("./user");
+const access = require("./access");
 
 class Job {
-    constructor(id, title, description, requirement, salaryMin, salaryMax, targetedPeople) {
+    constructor(id, title, description, requirement, salaryMin, salaryMax, targetedPeople, userId) {
         this.id = id;
         this.title = title;
         this.description = description;
@@ -14,11 +16,11 @@ class Job {
         this.salaryMin = salaryMin;
         this.salaryMax = salaryMax;
         this.targetedPeople = targetedPeople;
+        this.userId = userId;
     }
 }
 
-
-exports.addJob = (req, res, next) => {
+exports.addJob = async (req, res, next) => {
     const job = new Job(
         req.body.id,
         req.body.title,
@@ -26,18 +28,49 @@ exports.addJob = (req, res, next) => {
         req.body.requirement,
         req.body.salaryMin,
         req.body.salaryMax,
-        req.body.targetedPeople
+        req.body.targetedPeople,
+        req.body.userId
     );
-    auth.authenticateToken();
-    addData.addDataToTable(req, res, next, job);
+
+
+
+    if (await access.BlockSeeker(req, res, next) == true) {
+        addData.addDataToTable(req, res, next, job);
+    } else
+        next(new AppError('Not authorized', 401));
 };
 
-exports.updateJob = (req, res, next) => {
-    updateData.updateTable(req, res, next, "jobs");
+exports.updateJob = async (req, res, next) => {
+
+    if (await access.onlyId(req.params.id, req, res, next) == true) {
+        updateData.updateTable(req, res, next, "jobs");
+    } else
+        next(new AppError('Not authorized', 401));
 };
 
-exports.deleteJob = (req, res, next) => {
-    deleteData.deleteDataFromTable(req, res, next, "jobs", req.params.id);
+
+exports.getJobById = (id) => {
+    const sql = 'SELECT * FROM jobs WHERE id = ?';
+    const values = [id];
+    return new Promise((resolve, reject) => {
+        conn.query(sql, values, (err, data) => {
+            if (err) {
+                reject(err);
+            } else {
+                resolve(data[0]);
+            }
+        });
+    });
+};
+
+exports.deleteJob = async (req, res, next) => {
+
+    const jobData = await this.getJobById(req.params.id);
+    console.log(jobData.userId);
+    if (await access.onlyId(jobData.userId, req, res, next) == true) {
+        deleteData.deleteDataFromTable(req, res, next, "jobs", req.params.id);
+    } else
+        next(new AppError('Not authorized', 401));
 };
